requirements.txt
Django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
packaging==21.0
psycopg2==2.9.1
requests==2.26.0
shortuuid==1.0.1
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
packaging==21.0
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
shortuuid==1.0.1
sqlparse==0.4.2
